+++ PaulEngine/src/PaulEngine/Renderer/Resource/Buffer.h
@@ -166,10 +166,10 @@ namespace PaulEngine {
 		virtual const BufferLayout& GetLayout() const = 0;
 		virtual void SetLayout(const BufferLayout& layout) = 0;
 
-		virtual void SetData(const void* data, uint32_t size, uint32_t offset = 0) = 0;
+		virtual void SetData(const void* data, size_t size, size_t offset = 0) = 0;
 
-		static Ref<VertexBuffer> Create(uint32_t size, BufferUsage usage = BufferUsage::DYNAMIC_DRAW);
-		static Ref<VertexBuffer> Create(float* vertices, uint32_t size, BufferUsage usage = BufferUsage::STATIC_DRAW);
+		static Ref<VertexBuffer> Create(size_t size, BufferUsage usage = BufferUsage::DYNAMIC_DRAW);
+		static Ref<VertexBuffer> Create(float* vertices, size_t size, BufferUsage usage = BufferUsage::STATIC_DRAW);
 	};
 
 	class IndexBuffer
@@ -182,7 +182,7 @@ namespace PaulEngine {
 
 		inline virtual uint32_t GetCount() = 0;
 
-		virtual void SetData(const uint32_t* indices, uint32_t count, uint32_t offset = 0) = 0;
+		virtual void SetData(const uint32_t* indices, uint32_t count, size_t offset = 0) = 0;
 
 		static Ref<IndexBuffer> Create(uint32_t count, BufferUsage usage = BufferUsage::DYNAMIC_DRAW);
 		static Ref<IndexBuffer> Create(uint32_t* indices, uint32_t count, BufferUsage usage = BufferUsage::STATIC_DRAW);
